[IMP] lcc_members: new portal structures data management
