+++ lcc_members/controllers/__init__.py
@@ -1,4 +1,4 @@
 # Copyright 2020 Lokavaluto ()
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import portal_partner_profile
+from . import portal_structure_profile
